0-3_AI_Evaluation_Engine/설계문서_V6.0/requirements.txt
anthropic==0.40.0
supabase==2.3.4
python-dotenv==1.0.0
